+++ docs/app/Components/Sidebar/Sidebar.js
@@ -1,36 +1,41 @@
 import _ from 'lodash/fp'
-import React, { Component } from 'react'
+import React, { Component, PropTypes } from 'react'
+import { Link } from 'react-router'
 import * as stardust from 'stardust'
+import { typeOrder } from 'docs/app/utils'
 import META from 'src/utils/Meta'
 
-const { Menu, Input } = stardust
+const { Menu, Icon, Input } = stardust
 
 export default class Sidebar extends Component {
+  static propTypes = {
+    style: PropTypes.object,
+  }
   state = { query: '' }
 
   handleSearchChange = e => this.setState({ query: e.target.value })
 
-  getComponentsByQuery() {
-    return _.filter(component => {
-      const name = component._meta.name
-      const isParent = META.isParent(component)
-      const isQueryMatch = new RegExp(this.state.query, 'i').test(name)
-      return isParent && isQueryMatch
-    }, stardust)
-  }
-
-  getComponentsByType = type => {
+  renderItemsByType = (type) => {
     const items = _.flow(
-      _.filter(component => META.isType(component, type)),
-      _.sortBy((component, name) => name),
-      _.map(component => {
-        const name = component._meta.name
-        return <Menu.Item key={name} name={name} href={`#${name}`} />
+      _.filter(_.overEvery([
+        META.isParent,
+        META.isType(type),
+        ({ _meta }) => new RegExp(this.state.query, 'i').test(_meta.name),
+      ])),
+      _.sortBy('_meta.name'),
+      _.map(({ _meta }) => {
+        const route = `${_meta.type}s/${_.kebabCase(_meta.name)}`
+
+        return (
+          <Link to={route} className='item' activeClassName='active' key={_meta.name}>
+            {_meta.name}
+          </Link>
+        )
       })
-    )(this.getComponentsByQuery())
+    )(stardust)
 
     return _.isEmpty(items) ? [] : (
-      <div className='item'>
+      <div className='item' key={type}>
         <div className='header'>{_.capitalize(type)}s</div>
         <div className='menu'>{items}</div>
       </div>
@@ -38,15 +43,17 @@ export default class Sidebar extends Component {
   }
 
   render() {
-    const addons = this.getComponentsByType(META.type.addon)
-    const elements = this.getComponentsByType(META.type.element)
-    const collections = this.getComponentsByType(META.type.collection)
-    const views = this.getComponentsByType(META.type.view)
-    const modules = this.getComponentsByType(META.type.module)
-
+    const { style } = this.props
     return (
-      <Menu className='inverted secondary vertical fluid' style={{ margin: 0 }}>
-        <Menu.Item>
+      <Menu className='vertical fixed inverted' style={{ ...style }}>
+        <div className='item'>
+          <img src='http://semantic-ui.com/images/logo.png' style={{ marginRight: '1em' }} />
+          <strong>UI-React Docs</strong>
+        </div>
+        <a className='item' href='https://github.com/TechnologyAdvice/stardust'>
+          <Icon className='github' /> GitHub
+        </a>
+        <div className='item'>
           <Input
             className='transparent inverted icon'
             icon='search'
@@ -54,12 +61,11 @@ export default class Sidebar extends Component {
             iconClass='search link icon'
             onChange={this.handleSearchChange}
           />
-        </Menu.Item>
-        {addons}
-        {elements}
-        {collections}
-        {views}
-        {modules}
+        </div>
+        {_.map(this.renderItemsByType, typeOrder)}
+        <a className='item' href='https://github.com/TechnologyAdvice/stardust/blob/master/CHANGELOG.md'>
+          <Icon className='file text outline' /> CHANGELOG
+        </a>
       </Menu>
     )
   }