+++ docs/app/Examples/addons/Radio/Types/Toggle.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import { Radio } from 'stardust'
+
+const RadioToggleExample = () => (
+  <Radio type='toggle' />
+)
+
+export default RadioToggleExample