+++ docs/app/Examples/collections/Breadcrumb/Content/BreadcrumbSectionPropExample.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Breadcrumb } from 'stardust'
+
+const sections = [
+  { text: 'Home', link: true },
+  { text: 'Search', active: true },
+]
+
+const BreadcrumbSectionPropExample = () => (
+  <Breadcrumb sections={sections} />
+)
+
+export default BreadcrumbSectionPropExample