+++ docs/app/Examples/collections/Table/index.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react'
+import Variations from './Variations'
+
+export default class TableExamples extends Component {
+  render() {
+    return (
+      <div>
+        <Variations />
+      </div>
+    )
+  }
+}