+++ docs/app/Examples/elements/Loader/States/LoaderActiveExample.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import { Loader, Image, Segment } from 'stardust'
+
+const LoaderActiveExample = () => (
+  <Segment>
+    <Loader active />
+
+    <Image src='http://semantic-ui.com/images/wireframe/short-paragraph.png' />
+  </Segment>
+)
+
+export default LoaderActiveExample